Added default values and renamed mockData
