+++ src/pages/CompanyCourseManagement/CompanyCourseManagementPage.jsx
@@ -1,8 +1,8 @@
-import { Button, Grid } from '@mui/material';
-import PageLayout from '../../components/PageLayout';
-import { v4 as uuidv4 } from 'uuid';
-import mockCourses from './mockCourses';
-import CourseListItem from './components/CourseListItem';
+import { Button, Grid } from "@mui/material";
+import PageLayout from "../../components/PageLayout";
+import { v4 as uuidv4 } from "uuid";
+import mockCourses from "./mockData";
+import CourseListItem from "./components/CourseListItem";
 
 /**
  * Renders the Company Course Management Page.
@@ -10,17 +10,17 @@ import CourseListItem from './components/CourseListItem';
  */
 function CompanyCourseManagementPage() {
   const handleAddCourseClick = () => {
-    console.log('Add Course button is clicked');
+    console.log("Add Course button is clicked");
   };
 
   return (
-    <PageLayout title='Company Course Management'>
+    <PageLayout title="Company Course Management">
       <Button
-        variant='contained'
-        color='primary'
-        size='large'
+        variant="contained"
+        color="primary"
+        size="large"
         sx={{
-          margin: '116px 0 0px 116px',
+          margin: "116px 0 0px 116px",
         }}
         onClick={handleAddCourseClick}
       >
@@ -33,13 +33,13 @@ function CompanyCourseManagementPage() {
         columnSpacing={10}
         columns={16}
         sx={{
-          marginTop: '50px',
-          display: 'flex',
-          justifyContent: 'center',
-          marginBottom: '50px',
+          marginTop: "50px",
+          display: "flex",
+          justifyContent: "center",
+          marginBottom: "50px",
         }}
       >
-        {mockCourses.map((course) => (
+        {mockCourses?.map((course) => (
           <CourseListItem
             key={uuidv4()}
             courseName={course.courseName}
