+++ src/pages/CompanyTrainingsOverview/CompanyTrainingsOverviewPage.jsx
@@ -1,8 +1,8 @@
-import { Button, Grid } from '@mui/material';
-import PageLayout from '../../components/PageLayout';
-import { v4 as uuidv4 } from 'uuid';
-import mockTrainings from './mockTrainings';
-import TrainingListItem from './components/TrainingListItem';
+import { Button, Grid } from "@mui/material";
+import PageLayout from "../../components/PageLayout";
+import { v4 as uuidv4 } from "uuid";
+import mockTrainings from "./mockData";
+import TrainingListItem from "./components/TrainingListItem";
 
 /**
  * Renders the Company Trainings Overview Page.
@@ -10,17 +10,17 @@ import TrainingListItem from './components/TrainingListItem';
  */
 function CompanyTrainingsOverviewPage() {
   const handleAddTrainingClick = () => {
-    console.log('Add Training button is clicked');
+    console.log("Add Training button is clicked");
   };
 
   return (
-    <PageLayout title='Company Trainings Overview'>
+    <PageLayout title="Company Trainings Overview">
       <Button
-        variant='contained'
-        color='primary'
-        size='large'
+        variant="contained"
+        color="primary"
+        size="large"
         sx={{
-          margin: '116px 0 0px 116px',
+          margin: "116px 0 0px 116px",
         }}
         onClick={handleAddTrainingClick}
       >
@@ -33,13 +33,13 @@ function CompanyTrainingsOverviewPage() {
         columnSpacing={10}
         columns={16}
         sx={{
-          marginTop: '50px',
-          display: 'flex',
-          justifyContent: 'center',
-          marginBottom: '50px',
+          marginTop: "50px",
+          display: "flex",
+          justifyContent: "center",
+          marginBottom: "50px",
         }}
       >
-        {mockTrainings.map((training) => (
+        {mockTrainings?.map((training) => (
           <TrainingListItem
             key={uuidv4()}
             trainingName={training.trainingName}
